pending/tablet_assistant/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "tablet-assistant"
version = "0.1.0"
description = "AI-powered tablet overlay assistant"
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "httpx>=0.27.0",
    "pillow>=10.0",
    "opencv-python>=4.8.0",
    "numpy>=1.24.0",
    "adb-shell>=0.4.0",
    "websockets>=12.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.0", "black>=23.0", "ruff>=0.1.0"]
